Sort input
